Fixed errors with logging
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@1439 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserverweb/source/http_file.cpp
@@ -273,7 +273,7 @@ int HttpFile::send(HttpThreadContext* td, ConnectionPtr s, const char *filenameP
         {
           if(!td->appendOutputs)
           {
-            ret=(u_long)s->socket.send(td->buffer->GetBuffer(), nbr, 0);
+            ret=s->socket.send(td->buffer->GetBuffer(), nbr, 0);
             if(ret==SOCKET_ERROR)
             {
               ostringstream buffer;
